Move api_auth() out to a file that can be included from plugins/modules to allow them to expose their own programmatic API.
